Simple Asp Core Web Api Project
The project has been completed and the unit test has been written for
the services.

I had a problem saving the authentication session where the system is
not authorized after generating the jwt token. Unfortunately, I could
not solve the problem of this part.

You can see the description of the package from the document.txt file

Thank you Mr.Ghamsari.